refactoring - move marked folder to markdown (refs #12)
